pyproject.toml
[project]
name = "cyberjournal"
version = "0.1.0"
description = "Terminal-first encrypted personal journal with retro TUI"
requires-python = ">=3.12"
dependencies = [
    "textual",
    "rich",
    "aiosqlite",
    "cryptography",
    "argon2-cffi",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
